% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnpdfRox.R
\name{mvnpdf}
\alias{mvnpdf}
\title{Multivariate-Normal probability density function}
\usage{
mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)
}
\arguments{
\item{x}{p x n data matrix with n the number of observations and
p the number of dimensions}

\item{mean}{mean vector or list of mean vectors (either a vector or
a matrix)}

\item{varcovM}{variance-covariance matrix or list of variance-covariance
matrices (either a matrix or a list)}

\item{Log}{logical flag for returning the log of the probability density
function. Default is \code{TRUE}.}
}
\value{
a list containing the input matrix x and y the multivariate-Normal probability density function
computed at x
}
\description{
This is a concise description of what the function does.
}
\details{
This part gives more details on the function.
}
\examples{
mvnpdf(x=matrix(1.96), Log=FALSE)
dnorm(1.96)

mvnpdf(x=matrix(rep(1.96, 2), nrow=2, ncol=1), Log=FALSE)
}
